feat: 객체의 책임을 단순화하기 위해 Cars와 Winners 클래스 분리
