+++ src/main/java/racingcar/model/Cars.java
@@ -8,7 +8,6 @@ public class Cars {
 
     private static final String NAME_DELIMITER = ",";
     private List<Car> carList = new ArrayList<>();
-    ;
     Validator validator = new Validator();
 
     public Cars(String names) {
@@ -24,34 +23,10 @@ public class Cars {
         return carList;
     }
 
-    public List<String> getWinners() {
-        List<String> winnerList = calculateWinners();
-        return winnerList;
-    }
-
     public void moveCars() {
         for (Car car : carList) {
             car.setMovement();
         }
-    }
-
-    private List<String> calculateWinners() {
-        List<String> winnerList = new ArrayList<>();
-        int max = findMaxMovement();
-        for (Car car : carList) {
-            if (car.getMovement() == max) {
-                winnerList.add(car.getName());
-            }
-        }
-        return winnerList;
-    }
-
-    private int findMaxMovement() {
-        int max = Integer.MIN_VALUE;
-        for (Car car : carList) {
-            max = Math.max(max, car.getMovement());
-        }
-        return max;
     }
 
     private void nameListValidate(List<String> nameList) {
